Re-add mangling of uppercased Go field names to lowercase, fix issues preventing nested bindings from working.
this means in html, any leading uppercase characters in a go variable name should be lowercased. This is necessary due to polymer's poor handling of fields starting with an uppercase in certain situations (specifically, when matching them with a case insensitive attribute setting it.

Additionally, the fixes for the nesting issues mean that the polymer Go bindings will only push non empty Go values to JS land on creation, this is so Polymer doesn't overwrite the filled value on a parent because the child's empty binding got a notification